Reorganized components that are intentionally meant to override the ones from the starter theme
