+++ web/woatw-20191224/components/hero-image.js
