+++ web/woatw-20191224/components/img-wrapper.js
