+++ web/woatw-20191224/content/artist-landing-page.mdx
@@ -1,8 +1,8 @@
-import ShowsV2 from '../components/fix-for-locale-display-bug-in-theme/shows.js'
-import ReleasesV2 from '../components/fix-for-locale-display-bug-in-theme/releases.js'
+import ShowsV2 from '../components/shows.js'
+import ReleasesV2 from '../components/releases.js'
 import Subscribe from '../src/components/mailchimp/subscribe.js'
-import LandingSectionTitle from '../components/fix-for-locale-display-bug-in-theme/landing-section-title.js'
-import BannerV2 from '../components/fix-for-locale-display-bug-in-theme/banner.js'
+import LandingSectionTitle from '../components/landing-section-title.js'
+import BannerV2 from '../components/banner.js'
 
 <BannerV2 />
